Add a 'modification' java package for entities and add equipment id for branch status modification (#67)
