+++ src/main/java/org/gridsuite/modification/server/entities/equipment/modification/EquipmentModificationEntity.java
@@ -4,12 +4,13 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.modification.server.entities;
+package org.gridsuite.modification.server.entities.equipment.modification;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.gridsuite.modification.server.ModificationType;
 import org.gridsuite.modification.server.dto.EquipmenModificationInfos;
+import org.gridsuite.modification.server.entities.ModificationEntity;
 
 import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
