+++ src/main/java/org/gridsuite/modification/server/entities/equipment/modification/attribute/BooleanEquipmentAttributeModificationEntity.java
@@ -4,7 +4,7 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.modification.server.entities.equipment.attribute.modification;
+package org.gridsuite.modification.server.entities.equipment.modification.attribute;
 
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
