+++ src/main/java/org/gridsuite/modification/server/entities/equipment/modification/attribute/FloatEquipmentAttributeModificationEntity.java
@@ -4,7 +4,7 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.modification.server.entities.equipment.attribute.modification;
+package org.gridsuite.modification.server.entities.equipment.modification.attribute;
 
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
@@ -20,10 +20,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 @Entity
-@Table(name = "doubleEquipmentAttributeModification")
-@PrimaryKeyJoinColumn(foreignKey = @ForeignKey(name = "double_equipment_attribute_modification_id_fk_constraint"))
-public class DoubleEquipmentAttributeModificationEntity extends EquipmentAttributeModificationEntity<Double> {
-    public DoubleEquipmentAttributeModificationEntity(String equipmentId, String attributeName, Double attributeValue) {
+@Table(name = "floatEquipmentAttributeModification")
+@PrimaryKeyJoinColumn(foreignKey = @ForeignKey(name = "float_equipment_attribute_modification_id_fk_constraint"))
+public class FloatEquipmentAttributeModificationEntity extends EquipmentAttributeModificationEntity<Float> {
+    public FloatEquipmentAttributeModificationEntity(String equipmentId, String attributeName, Float attributeValue) {
         super(equipmentId, attributeName, attributeValue);
     }
 }
