+++ src/main/resources/db/changelog/changesets/changelog_2021-12-28T09:12:53Z.xml
@@ -1,26 +1,15 @@
 <?xml version="1.1" encoding="UTF-8" standalone="no"?>
 <databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog" xmlns:ext="http://www.liquibase.org/xml/ns/dbchangelog-ext" xmlns:pro="http://www.liquibase.org/xml/ns/pro" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd http://www.liquibase.org/xml/ns/pro http://www.liquibase.org/xml/ns/pro/liquibase-pro-3.10.xsd http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.10.xsd">
-    <changeSet author="amarsli (generated)" id="1640357270577-5">
+    <changeSet author="amarsli (generated)" id="1640682785313-5">
         <createTable tableName="branch_status_modification">
+            <column name="equipment_id" type="VARCHAR(255)"/>
             <column name="action" type="VARCHAR(255)"/>
             <column name="id" type="UUID">
                 <constraints nullable="false" primaryKey="true" primaryKeyName="branch_status_modificationPK"/>
             </column>
         </createTable>
     </changeSet>
-    <changeSet author="amarsli (generated)" id="1640357270577-6">
+    <changeSet author="amarsli (generated)" id="1640682785313-6">
         <addForeignKeyConstraint baseColumnNames="id" baseTableName="branch_status_modification" constraintName="branchStatusModification_id_fk_constraint" deferrable="false" initiallyDeferred="false" referencedColumnNames="id" referencedTableName="modification" validate="true"/>
-    </changeSet>
-    <changeSet author="amarsli (generated)" id="1640357270577-1">
-        <dropForeignKeyConstraint baseTableName="two_windings_transformer_creation" constraintName="current_limits_id1_fk"/>
-    </changeSet>
-    <changeSet author="amarsli (generated)" id="1640357270577-2">
-        <addForeignKeyConstraint baseColumnNames="current_limits_id1" baseTableName="two_windings_transformer_creation" constraintName="current_limits_id1_fk" referencedColumnNames="id" referencedTableName="current_limits"/>
-    </changeSet>
-    <changeSet author="amarsli (generated)" id="1640357270577-3">
-        <dropForeignKeyConstraint baseTableName="two_windings_transformer_creation" constraintName="current_limits_id2_fk"/>
-    </changeSet>
-    <changeSet author="amarsli (generated)" id="1640357270577-4">
-        <addForeignKeyConstraint baseColumnNames="current_limits_id2" baseTableName="two_windings_transformer_creation" constraintName="current_limits_id2_fk" referencedColumnNames="id" referencedTableName="current_limits"/>
     </changeSet>
 </databaseChangeLog>
