+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -16,5 +16,5 @@ databaseChangeLog:
       file: changesets/changelog_2021-12-22T12:16:24Z.xml
       relativeToChangelogFile: true
   - include:
-      file: changesets/changelog_2021-12-24T14:47:39Z.xml
+      file: changesets/changelog_2021-12-28T09:12:53Z.xml
       relativeToChangelogFile: true
